check if taxonomy exists on add new Advanced tab

+++ inc/class-admin-add-item.php
@@ -203,6 +203,8 @@ class SellMediaAdminAddItem {
 		echo '<div id="sell-media-advanced-options-container">';
 
 		// do_action( 'sell_media_add_item_tab_before_content_callback', 'advanced', $post );
+		
+		if ( taxonomy_exists( 'collection' ) ) {
 
 			echo '<div id="sell-media-tax-collections" class="sell-media-tax-wrap">';
 				printf( '<h3 class="tax-title">%s</h3>', esc_html__( 'Collections', 'sell_media' ) );
@@ -211,12 +213,20 @@ class SellMediaAdminAddItem {
 				printf( '<div class="tax-edit"><a href="' . esc_url( admin_url( 'edit-tags.php?taxonomy=collection&post_type=sell_media_item' ) ) . '">%s</a></div>', esc_html__( 'Edit All Collections', 'sell_media' ) );
 			echo '</div>';
 
+		}
+
+		if ( taxonomy_exists( 'licenses' ) ) {
+
 			echo '<div id="sell-media-tax-licenses" class="sell-media-tax-wrap">';
 				printf( '<h3 class="tax-title">%s</h3>', esc_html__( 'Licenses', 'sell_media' ) );
 				printf( '<p class="tax-description description">%s.</p>', esc_html__( 'Select the available usage licenses that buyers can choose from when purchasing (optional). Licenses can be assigned "markup" which will increase the cost of the item being purchase. For example, you can might have a "Personal" usage license with no markup from your base pricelists and a "Commercial" usage license with 50% markup from your base pricelists', 'sell_media' ) );
 				post_categories_meta_box( $post, array( 'args' => array( 'taxonomy' => 'licenses' ) ) );
 				printf( '<div class="tax-edit"><a href="' . esc_url( admin_url( 'edit-tags.php?taxonomy=licenses&post_type=sell_media_item' ) ) . '">%s</a></div>', esc_html__( 'Edit All Licenses', 'sell_media' ) );
 			echo '</div>';
+
+		}
+
+		if ( taxonomy_exists( 'creator' ) ) {
 
 			echo '<div id="sell-media-tax-creators" class="sell-media-tax-wrap">';
 				printf( '<h3 class="tax-title">%s</h3>', esc_html__( 'Creators', 'sell_media' ) );
@@ -225,6 +235,7 @@ class SellMediaAdminAddItem {
 				// post_tags_meta_box( $post, array( 'args' => array( 'taxonomy' => 'creator' ) ) );
 				printf( '<div class="tax-edit"><a href="' . esc_url( admin_url( 'edit-tags.php?taxonomy=creator&post_type=sell_media_item' ) ) . '">%s</a></div>', esc_html__( 'Edit All Creators', 'sell_media' ) );
 			echo '</div>';
+		}
 
 		echo '</div>';
 	}
